database: Move `OauthGithub` model into dedicated module
This moves `OauthGithub`, `NewOauthGithub`, and its `insert()` method
from `user.rs` into `oauth_github.rs`, keeping the GitHub account models
together in their own file. The existing `models` re-exports and
database behavior remain unchanged.
