+++ crates/crates_io_database/src/models/mod.rs
@@ -14,11 +14,12 @@ pub use self::email::{Email, NewEmail};
 pub use self::follow::Follow;
 pub use self::keyword::{CrateKeyword, Keyword};
 pub use self::krate::{Crate, CrateName, NewCrate};
+pub use self::oauth_github::{NewOauthGithub, OauthGithub};
 pub use self::owner::{CrateOwner, Owner, OwnerKind};
 pub use self::team::{NewTeam, Team};
 pub use self::token::ApiToken;
 pub use self::trustpub::TrustpubData;
-pub use self::user::{NewOauthGithub, NewUser, OauthGithub, PublicUser, User, users_by_username};
+pub use self::user::{NewUser, PublicUser, User, users_by_username};
 pub use self::version::{NewVersion, TopVersions, Version};
 
 pub mod helpers;
@@ -35,6 +36,7 @@ mod email;
 mod follow;
 mod keyword;
 pub mod krate;
+mod oauth_github;
 mod owner;
 pub mod team;
 pub mod token;
